pyproject.toml
[project]
name = "doc-demo"
version = "0.1.0"
description = "LangGraph + Gradio prototype for python-docx document workflows"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gradio>=6.9.0",
    "langchain>=1.2.12",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.2",
    "pydantic>=2.12.5",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
]

[project.scripts]
doc-demo = "doc_demo.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/doc_demo"]
